Separate PluginService tests
